Move components/Group -> views/components/CreateGroupSteps
