+++ frontend/simple/main.js
@@ -7,7 +7,7 @@ import Vue from 'vue'
 import router from './controller/router'
 import './js/translations'
 import * as db from './model/database'
-import NavBar from './views/views/NavBar.vue'
+import NavBar from './views/containers/NavBar.vue'
 import './js/transitions'
 import {namespace} from './controller/backend/hapi'
 import store from './model/state'
