+++ frontend/simple/views/components/CreateGroupSteps/GroupPrivacy.vue
