+++ frontend/simple/views/components/CreateGroupSteps/index.js
