+++ frontend/simple/views/views/Invite.vue
@@ -47,7 +47,7 @@ import * as Events from '../../../../shared/events'
 import backend from '../../controller/backend/'
 import L from '../../js/translations'
 import template from 'string-template'
-import { GroupInvitees } from '../../components/Group'
+import { GroupInvitees } from '../components/CreateGroupSteps'
 import { mapState, mapGetters } from 'vuex'
 import sbp from '../../../../shared/sbp'
