+++ shared/declarations.js
@@ -101,7 +101,7 @@ declare module './GroupPrivacy.vue' { declare module.exports: Object }
 declare module './GroupInvitees.vue' { declare module.exports: Object }
 declare module './GroupSummary.vue' { declare module.exports: Object }
 declare module '../views/components/i18n.vue' { declare module.exports: Object }
-declare module './views/views/NavBar.vue' { declare module.exports: Object }
+declare module './views/containers/NavBar.vue' { declare module.exports: Object }
 declare module '../views/views/CreateGroup.vue' { declare module.exports: Object }
 declare module '../views/views/PayGroup.vue' { declare module.exports: Object }
 declare module '../views/views/SignUp.vue' { declare module.exports: Object }
